requirements.txt
asyncclick
emoji
python-telegram-bot
